delivery-map-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
clickhouse-connect==0.6.23
python-dotenv==1.0.0
python-multipart==0.0.6
openpyxl==3.1.5
gspread==5.12.4
google-auth==2.23.4
httpx==0.28.1
pandas
